+++ cognite/experimental/data_classes/geospatial.py
@@ -1,6 +1,7 @@
-from typing import Dict
+from typing import Any, Dict, List
 
 from cognite.client import utils
+from cognite.client.data_classes._base import CogniteResource, CogniteResourceList
 
 
 class RasterMetadata:
@@ -17,3 +18,30 @@ class RasterMetadata:
             snake_case_key = utils._auxiliary.to_snake_case(key)
             setattr(instance, snake_case_key, value)
         return instance
+
+
+class MvpMappingsDefinition(CogniteResource):
+    """MVT mappings definition"""
+
+    def __init__(
+        self,
+        external_id: str = None,
+        mappings: List[Dict[str, Any]] = None,
+        cognite_client=None,
+    ):
+        self.external_id = external_id
+        self.mappings = mappings
+        self._cognite_client = cognite_client
+
+    @classmethod
+    def _load(cls, resource: Dict, cognite_client=None):
+        instance = cls(cognite_client=cognite_client)
+        for key, value in resource.items():
+            snake_case_key = utils._auxiliary.to_snake_case(key)
+            setattr(instance, snake_case_key, value)
+        return instance
+
+
+class MvpMappingsDefinitionList(CogniteResourceList):
+    _RESOURCE = MvpMappingsDefinition
+    _ASSERT_CLASSES = False