+++ tests/tests_integration/test_api/test_geospatial.py
@@ -4,6 +4,7 @@ import pytest
 from cognite.client.data_classes.geospatial import *
 
 from cognite.experimental import CogniteClient
+from cognite.experimental.data_classes.geospatial import *
 
 
 @pytest.fixture(scope="class")
@@ -39,7 +40,7 @@ def test_feature_type(cognite_client, cognite_domain):
 
 @pytest.fixture
 def test_feature(cognite_client, test_feature_type):
-    external_id = f"F_{uuid.uuid4().hex[:10]}"
+    external_id = f"F/O@ö_{uuid.uuid4().hex[:10]}"
     feature = cognite_client.geospatial.create_features(
         test_feature_type.external_id,
         Feature(
@@ -59,7 +60,7 @@ def test_feature_with_raster(cognite_client, test_feature_type, test_feature):
     cognite_client.geospatial.put_raster(
         feature_type_external_id=test_feature_type.external_id,
         feature_external_id=test_feature.external_id,
-        raster_id="raster",
+        raster_property_name="raster",
         raster_format="XYZ",
         raster_srid=3857,
         file="tests/tests_integration/test_api/geospatial_data/raster-grid-example.xyz",
@@ -67,10 +68,27 @@ def test_feature_with_raster(cognite_client, test_feature_type, test_feature):
     yield test_feature
 
 
-# NB: raster tests are for now marked as skip since GCP has GDAL drivers entirely disable.
-# There is no way to enable them yet, so testing can be done manually against azure-dev or bluefield
-# by commenting out the @pytest.mark.skip annotation
-class TestGeospatialAPI:
+@pytest.fixture
+def test_mvt_mappings_def(cognite_client, test_feature_type):
+    external_id = f"MVT/O@ö_{uuid.uuid4().hex[:10]}"
+    mvt_mappings_def = cognite_client.geospatial.create_mvt_mappings_definitions(
+        MvpMappingsDefinition(
+            external_id=external_id,
+            mappings=[
+                {
+                    "featureTypeExternalId": test_feature_type.external_id,
+                    "levels": [0, 1, 2, 3, 4, 5],
+                    "geometryProperty": "position",
+                    "featureProperties": ["volume", "temperature"],
+                }
+            ],
+        )
+    )
+    yield mvt_mappings_def
+    cognite_client.geospatial.delete_mvt_mappings_definitions(external_id=external_id)
+
+
+class TestExperimentalGeospatialAPI:
 
     # This test already exist in the main python sdk
     # It is repeated here to test the geospatial domain part.
@@ -99,27 +117,27 @@ class TestGeospatialAPI:
         res = cognite_client.geospatial.put_raster(
             feature_type_external_id=test_feature_type.external_id,
             feature_external_id=test_feature.external_id,
-            raster_id="raster",
+            raster_property_name="raster",
             raster_format="XYZ",
             raster_srid=3857,
             file="tests/tests_integration/test_api/geospatial_data/raster-grid-example.xyz",
         )
         assert res.width == 4
         assert res.height == 5
-        assert res.numbands == 1
-        assert res.scalex == 1.0
-        assert res.scaley == 1.0
-        assert res.skewx == 0.0
-        assert res.skewy == 0.0
+        assert res.num_bands == 1
+        assert res.scale_x == 1.0
+        assert res.scale_y == 1.0
+        assert res.skew_x == 0.0
+        assert res.skew_y == 0.0
         assert res.srid == 3857
-        assert res.upperleftx == -0.5
-        assert res.upperlefty == -0.5
+        assert res.upper_left_x == -0.5
+        assert res.upper_left_y == -0.5
 
     def test_get_raster(self, cognite_client, test_feature_type, test_feature_with_raster):
         res = cognite_client.geospatial.get_raster(
             feature_type_external_id=test_feature_type.external_id,
             feature_external_id=test_feature_with_raster.external_id,
-            raster_id="raster",
+            raster_property_name="raster",
             raster_format="XYZ",
         )
         raster_content = open("tests/tests_integration/test_api/geospatial_data/raster-grid-example.xyz", "rb").read()
@@ -128,7 +146,7 @@ class TestGeospatialAPI:
         res = cognite_client.geospatial.get_raster(
             feature_type_external_id=test_feature_type.external_id,
             feature_external_id=test_feature_with_raster.external_id,
-            raster_id="raster",
+            raster_property_name="raster",
             raster_format="XYZ",
             raster_options={"ADD_HEADER_LINE": "YES"},
         )
@@ -137,34 +155,71 @@ class TestGeospatialAPI:
         ).read()
         assert res == raster_content
 
+    def test_get_raster_with_transformation(self, cognite_client, test_feature_type, test_feature_with_raster):
+        res = cognite_client.geospatial.get_raster(
+            feature_type_external_id=test_feature_type.external_id,
+            feature_external_id=test_feature_with_raster.external_id,
+            raster_property_name="raster",
+            raster_format="XYZ",
+            raster_srid=54030,
+            allow_crs_transformation=True,
+        )
+        raster_content = open(
+            "tests/tests_integration/test_api/geospatial_data/raster-grid-54030-example.xyz", "rb"
+        ).read()
+        assert res == raster_content
+
     def test_retrieve_features_with_raster_property(self, cognite_client, test_feature_type, test_feature_with_raster):
         res = cognite_client.geospatial.retrieve_features(
-            feature_type_external_id=test_feature_type.external_id, external_id=[test_feature_with_raster.external_id],
+            feature_type_external_id=test_feature_type.external_id,
+            external_id=[test_feature_with_raster.external_id],
         )
         assert res[0].external_id == test_feature_with_raster.external_id
         raster_metadata = res[0].raster
         assert raster_metadata == {
             "width": 4,
             "height": 5,
-            "numbands": 1,
-            "scalex": 1.0,
-            "scaley": 1.0,
-            "skewx": 0.0,
-            "skewy": 0.0,
+            "numBands": 1,
+            "scaleX": 1.0,
+            "scaleY": 1.0,
+            "skewX": 0.0,
+            "skewY": 0.0,
             "srid": 3857,
-            "upperleftx": -0.5,
-            "upperlefty": -0.5,
+            "upperLeftX": -0.5,
+            "upperLeftY": -0.5,
         }
+
+    def test_put_raster_custom_crs(self, cognite_client, test_feature_type, test_feature):
+        res = cognite_client.geospatial.put_raster(
+            feature_type_external_id=test_feature_type.external_id,
+            feature_external_id=test_feature.external_id,
+            raster_property_name="raster",
+            raster_format="XYZ",
+            raster_srid=54030,
+            file="tests/tests_integration/test_api/geospatial_data/raster-grid-example.xyz",
+            allow_crs_transformation=True,
+        )
+        assert res.width == 4
+        assert res.height == 5
+        assert res.num_bands == 1
+        assert res.scale_x == 1.0
+        assert res.scale_y == 1.0
+        assert res.skew_x == 0.0
+        assert res.skew_y == 0.0
+        assert res.srid == 3857
+        assert res.upper_left_x == -0.5891363261459447
+        assert res.upper_left_y == -0.31623471547260973
 
     def test_delete_raster(self, cognite_client, test_feature_type, test_feature_with_raster):
         res = cognite_client.geospatial.delete_raster(
             feature_type_external_id=test_feature_type.external_id,
             feature_external_id=test_feature_with_raster.external_id,
-            raster_id="raster",
+            raster_property_name="raster",
         )
         assert res is None
         res = cognite_client.geospatial.retrieve_features(
-            feature_type_external_id=test_feature_type.external_id, external_id=[test_feature_with_raster.external_id],
+            feature_type_external_id=test_feature_type.external_id,
+            external_id=[test_feature_with_raster.external_id],
         )
         assert res[0].external_id == test_feature_with_raster.external_id
         assert hasattr(res[0], "raster") is False
@@ -187,7 +242,20 @@ class TestGeospatialAPI:
             "position",
         }
         res = cognite_client.geospatial.retrieve_features(
-            feature_type_external_id=test_feature_type.external_id, external_id=[test_feature_with_raster.external_id],
+            feature_type_external_id=test_feature_type.external_id,
+            external_id=[test_feature_with_raster.external_id],
         )
         assert res[0].external_id == test_feature_with_raster.external_id
         assert hasattr(res[0], "raster") is False
+
+    def test_get_mvt_mappings_definitions(self, cognite_client, test_mvt_mappings_def):
+        res = cognite_client.geospatial.retrieve_mvt_mappings_definitions(external_id=test_mvt_mappings_def.external_id)
+        assert res.external_id == test_mvt_mappings_def.external_id
+        assert res.mappings[0]["levels"] == test_mvt_mappings_def.mappings[0]["levels"]
+        assert res.mappings[0]["geometryProperty"] == test_mvt_mappings_def.mappings[0]["geometryProperty"]
+        assert res.mappings[0]["featureProperties"] == test_mvt_mappings_def.mappings[0]["featureProperties"]
+
+    def test_list_mvt_mappings_definitions(self, cognite_client, test_mvt_mappings_def):
+        res = cognite_client.geospatial.list_mvt_mappings_definitions()
+        assert len(res) == 1
+        assert res[0].external_id == test_mvt_mappings_def.external_id